Version 0.0.1.177:
- test-results.txt aktualisiert und umgebaut
- Code in tasking eingefügt, der später create_vm86_ThreadTaskBase ersetzen soll (funktioniert noch nicht)
- Code vereinfacht und Styleänderungen

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@756 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/network/rtl8139_eeprom.c
@@ -16,7 +16,7 @@ int read_eeprom(long ioaddr, int location, int addr_len)
     outportb(ee_addr, EE_ENB);
 
     // Shift the read command bits out
-    for (int i = 4 + addr_len; i >= 0; i--) 
+    for (int i = 4 + addr_len; i >= 0; i--)
     {
         int dataval = (read_cmd & (1 << i)) ? EE_DATA_WRITE : 0;
         outportb(ee_addr, EE_ENB | dataval);
@@ -27,7 +27,7 @@ int read_eeprom(long ioaddr, int location, int addr_len)
     outportb(ee_addr, EE_ENB);
     eeprom_delay();
 
-    for (int i = 16; i > 0; i--) 
+    for (int8_t i = 16; i > 0; i--)
     {
         outportb(ee_addr, EE_ENB | EE_SHIFT_CLK);
         eeprom_delay();
